+++ test/unit/zebra_sgd_commands_test.dart
@@ -1,30 +1,13 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:zebrautil/models/print_enums.dart';
-import 'package:zebrautil/zebra_sgd_commands.dart';
+import 'package:zebrautil/internal/parser_util.dart';
 
 void main() {
-  group('ZebraSGDCommands', () {
-    test('isZPLData detects ZPL', () {
-      expect(ZebraSGDCommands.isZPLData('^XA'), isTrue);
-      expect(ZebraSGDCommands.isZPLData('foo'), isFalse);
-    });
-    test('isCPCLData detects CPCL', () {
-      expect(ZebraSGDCommands.isCPCLData('! 0'), isTrue);
-      expect(ZebraSGDCommands.isCPCLData('^XA'), isFalse);
-    });
-    test('detectDataLanguage returns correct language', () {
-      expect(
-          ZebraSGDCommands.detectDataLanguage('^XA'), equals(PrintFormat.zpl));
-      expect(
-          ZebraSGDCommands.detectDataLanguage('! 0'), equals(PrintFormat.cpcl));
-      expect(ZebraSGDCommands.detectDataLanguage('foo'), isNull);
-    });
+  group('ParserUtil', () {
     test('parseResponse extracts value from SGD response', () {
-      expect(ZebraSGDCommands.parseResponse('"foo" : "bar"'), equals('bar'));
-      expect(ZebraSGDCommands.parseResponse('"bar"'), equals('bar'));
-      expect(ZebraSGDCommands.parseResponse('baz'), equals('baz'));
-      expect(ZebraSGDCommands.parseResponse(''), isNull);
+      expect(ParserUtil.parseResponse('"foo" : "bar"'), equals('bar'));
+      expect(ParserUtil.parseResponse('"bar"'), equals('bar'));
+      expect(ParserUtil.parseResponse('baz'), equals('baz'));
+      expect(ParserUtil.parseResponse(''), isNull);
     });
-    
   });
 }